A class that produces a "meaningful" pathway from a set of points.
From a limited set of data-points in a 2D map, we can produce a vast amount of different paths.
This class produces a pathway for the taxxing app, taking as input a set of points in the map .
Essentially it orders the set of points in a way that it would be meaningful inside the context of the airport lanes.

Taxiing app is a Unity application (released on 16/5/21 in beta mode) 
that helps pilots taxiing their plane using a given path from the ground control, 
written in NATO format (an example of this path is: Y1 C I H A A11 21R).
These initials correspond in general to some (but not all) of the roads the plane is crossing during its course. 

Taxiing app was designed with rapid prototyping in mind.
Meaning that we wanted our end product to be a template which we will be able in the near future 
to tweek and create rapidly a new map for another airport and make corrections without causing errors.
For this reason taxiing app has been created in Unity where we can easily generate UI points 
and move them freely in space. 
The names we give in these key points are used in conjuction with the given NATO path to produce the final pathway which is depicted in the map.  
One of the problems we originally faced, was that each path is not defined only by its NATO coded path 
but also by the intermediate points who define long curves of any possible shape.


This class solves this problem. By giving the initials: NP + [name of the lane]
recognizes that for instance point 1, point 16, point 5, point 9 are in the same lane and stores them in a separate array in sequence (road).
Key role in order to find which is the right sequence of the points is to identify if this spesific array has X or Y orientation. In order to find orientation, we calculate the first and the last point and then the distance between them.
Then the data points placed in a spesific order according to the axis we have found.

*This class at this point demonstrates a small glimpse of the work I have done in order to build the taxiing app
 and cannot work as a standalone app.
In a future realease I will create a standalone class that can be used outside of the app's context.